+++ Source/Reflection/Shared/WinCopies.Reflection.DotNetDocBuilder.Shared/Entities.cs
@@ -5,8 +5,18 @@ using WinCopies.EntityFramework;
 
 namespace WinCopies.Reflection.DotNetDocBuilder
 {
+    public class DefaultDBEntity<T> : DefaultDBEntity<T, ulong> where T : IEntity
+    {
+        public DefaultDBEntity(DBEntityCollection<T> collection) : base(collection) { /* Left empty. */ }
+    }
+
+    public class DefaultNamedEntity2<T> : DefaultNamedEntity2<T, ulong> where T : IEntity
+    {
+        public DefaultNamedEntity2(DBEntityCollection<T> collection) : base(collection) { /* Left empty. */ }
+    }
+
     [Entity("docnamespace")]
-    public class Namespace : DefaultDBEntity<Namespace, ulong>
+    public class Namespace : DefaultDBEntity<Namespace>
     {
         private int _frameworkId;
         private string _name;
@@ -42,32 +52,40 @@ namespace WinCopies.Reflection.DotNetDocBuilder
     }
 
     [Entity("typetype")]
-    public class TypeType : DefaultNamedEntity2<TypeType, ulong>
+    public class TypeType : DefaultNamedEntity2<TypeType>
     {
         public TypeType(DBEntityCollection<TypeType> collection) : base(collection) { /* Left empty. */ }
     }
 
     [Entity("class_access_modifier")]
-    public class AccessModifier : DefaultNamedEntity2<AccessModifier, ulong>
+    public class AccessModifier : DefaultNamedEntity2<AccessModifier>
     {
         public AccessModifier(DBEntityCollection<AccessModifier> collection) : base(collection) { /* Left empty. */ }
     }
 
-    [Entity("doctype")]
-    public class Type : DefaultDBEntity<Type, ulong>
+    public class DocItem<T> : DefaultNamedEntity2<T> where T : IEntity
     {
-        private string _name;
+        private string _comment;
+        private string _commentRemarks;
+
+        [EntityProperty]
+        public string Comment { get => TryRefreshAndGet(ref _comment); set => _comment = value; }
+
+        [EntityProperty]
+        public string CommentRemarks { get => TryRefreshAndGet(ref _commentRemarks); set => _commentRemarks = value; }
+
+        public DocItem(DBEntityCollection<T> collection) : base(collection) { /* Left empty. */ }
+    }
+
+    [Entity("doctype")]
+    public class Type : DocItem<Type>
+    {
         private Namespace _namespace;
         private TypeType _typeType;
         private AccessModifier _accessModifier;
         private byte _genericTypeCount;
         private Class _parentType;
-        private string _comment;
-        private string _commentRemarks;
         private string _toStringValue;
-
-        [EntityProperty(IsPseudoId = true)]
-        public string Name { get => TryRefreshAndGet(ref _name); set => _name = value; }
 
         [EntityProperty(
 #if CS10
@@ -104,12 +122,6 @@ namespace WinCopies.Reflection.DotNetDocBuilder
 #endif
             ParentType
         { get => TryRefreshAndGet(ref _parentType); set => _parentType = value; }
-
-        [EntityProperty]
-        public string Comment { get => TryRefreshAndGet(ref _comment); set => _comment = value; }
-
-        [EntityProperty]
-        public string CommentRemarks { get => TryRefreshAndGet(ref _commentRemarks); set => _commentRemarks = value; }
 
         public Type(DBEntityCollection<Type> collection) : base(collection) { /* Left empty. */ }
 
@@ -184,12 +196,12 @@ namespace WinCopies.Reflection.DotNetDocBuilder
     }
 
     [Entity("enumunderlyingtype")]
-    public class UnderlyingType : DefaultNamedEntity2<UnderlyingType, ulong>
+    public class UnderlyingType : DefaultNamedEntity2<UnderlyingType>
     {
         public UnderlyingType(DBEntityCollection<UnderlyingType> collection) : base(collection) { /* Left empty. */ }
     }
 
-    public class TypeBase<T> : DefaultDBEntity<T, ulong> where T : IEntity
+    public class TypeBase<T> : DefaultDBEntity<T> where T : IEntity
     {
         private Type _type;
 
@@ -229,7 +241,7 @@ namespace WinCopies.Reflection.DotNetDocBuilder
         Type GetImplementedInterface();
     }
 
-    public abstract class InterfaceImplementation<T> : DefaultDBEntity<T, ulong>, IInterfaceImplementation where T : IEntity
+    public abstract class InterfaceImplementation<T> : DefaultDBEntity<T>, IInterfaceImplementation where T : IEntity
     {
         private Type _implementedInterface;
 
@@ -245,7 +257,7 @@ namespace WinCopies.Reflection.DotNetDocBuilder
     }
 
     [Entity("class_modifier")]
-    public class ClassModifier : DefaultNamedEntity2<ClassModifier, ulong>
+    public class ClassModifier : DefaultNamedEntity2<ClassModifier>
     {
         public ClassModifier(DBEntityCollection<ClassModifier> collection) : base(collection) { /* Left empty. */ }
     }
@@ -337,18 +349,18 @@ namespace WinCopies.Reflection.DotNetDocBuilder
     }
 
     [Entity("docgenerictypemodifier")]
-    public class GenericTypeModifier : DefaultNamedEntity2<GenericTypeModifier, ulong>
+    public class GenericTypeModifier : DefaultNamedEntity2<GenericTypeModifier>
     {
         public GenericTypeModifier(DBEntityCollection<GenericTypeModifier> collection) : base(collection) { /* Left empty. */ }
     }
 
     [Entity("docgenerictype")]
-    public class GenericType : DefaultNamedEntity<GenericType, ulong>
+    public class GenericType : DefaultNamedEntity2<GenericType>
     {
         private Type _type;
         private GenericTypeModifier _modifier;
 
-        [EntityProperty("DocType")]
+        [EntityProperty("DocType", IsPseudoId = true)]
         [ForeignKey]
         public Type Type { get => TryRefreshAndGet(ref _type); set => _type = value; }
 
@@ -357,5 +369,51 @@ namespace WinCopies.Reflection.DotNetDocBuilder
         public GenericTypeModifier Modifier { get => TryRefreshAndGet(ref _modifier); set => _modifier = value; }
 
         public GenericType(DBEntityCollection<GenericType> collection) : base(collection) { /* Left empty. */ }
+    }
+
+    [Entity("docmember")]
+    public class Member : DocItem<Member>
+    {
+        private Type _type;
+
+        [EntityProperty("TypeId", IsPseudoId = true)]
+        [ForeignKey]
+        public Type Type { get => TryRefreshAndGet(ref _type); set => _type = value; }
+
+        public Member(DBEntityCollection<Member> collection) : base(collection) { /* Left empty. */ }
+    }
+
+    public interface IConst : IEntity
+    {
+        object Value { get; set; }
+    }
+
+    public class Const<TEntity, TValue> : DefaultDBEntity<TEntity>, IConst where TEntity : IEntity
+    {
+        private TValue _value;
+        private Member _memberId;
+
+        [EntityProperty]
+        public TValue Value { get => TryRefreshAndGet(ref _value); set => _value = value; }
+
+        object IConst.Value { get => Value; set => Value = (TValue)value; }
+
+        [EntityProperty(IsPseudoId = true)]
+        [ForeignKey(RemoveAlso = true)]
+        public Member Member { get => TryRefreshAndGet(ref _memberId); set => _memberId = value; }
+
+        public Const(DBEntityCollection<TEntity> collection) : base(collection) { /* Left empty. */ }
+    }
+
+    [Entity("docconst16")]
+    public class Const16 : Const<Const16, ushort>
+    {
+        public Const16(DBEntityCollection<Const16> collection) : base(collection) { /* Left empty. */ }
+    }
+
+    [Entity("docconst64")]
+    public class Const64 : Const<Const64, ulong>
+    {
+        public Const64(DBEntityCollection<Const64> collection) : base(collection) { /* Left empty. */ }
     }
 }